+++ Utilities/vtkexodus2/expcon.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1994 Sandia Corporation. Under the terms of Contract
+ * Copyright (c) 2005 Sandia Corporation. Under the terms of Contract
  * DE-AC04-94AL85000 with Sandia Corporation, the U.S. Governement
  * retains certain rights in this software.
  * 
@@ -53,14 +53,14 @@
 *
 * revision history - 
 *
-*  $Id: expcon.c,v 1.1 2005/07/17 15:44:00 andy Exp $
+*  $Id: expcon.c,v 1.3 2008-04-22 21:14:53 david.cole Exp $
 *
 *****************************************************************************/
 
 #include "exodusII.h"
 #include "exodusII_int.h"
 #include <string.h>
-/*
+/*!
  * writes the names of the coordinate arrays to the database
  */
 
@@ -114,7 +114,7 @@ int ex_put_coord_names (int   exoid,
      start[1] = 0;
 
      count[0] = 1;
-     count[1] = strlen(coord_names[i]) + 1;
+     count[1] = (long)strlen(coord_names[i]) + 1;
 
      if (ncvarput (exoid, varid, start, count, (void*) coord_names[i]) == -1)
      {
